v0.14

+++ Victoria II Economy Analyzer/src/main/java/org/victoria2/tools/vic2sgea/entities/Country.java
@@ -169,7 +169,7 @@ public class Country extends EconomySubject implements Comparable<Country> {
 	}
 
 	public float getUnemploymentRateFactory() {
-		return (workforceFactory - employmentFactory) / (float) workforceFactory * 100;
+		return (float) (workforceFactory - employmentFactory) / (float) workforceFactory * 100;
 	}
 
 	public long getWorkforceRgo() {
